+++ src/AVPlayerPrivate.cpp
@@ -451,23 +451,23 @@ void AVPlayer::Private::updateBufferValue(PacketBuffer* buf)
 {
     const bool video = vthread && buf == vthread->packetQueue();
     const qreal fps = qMax<qreal>(24.0, statistics.video.frame_rate);
-    int bv = 0.5*fps;
+    qint64 bv = 0.5*fps;
     if (!video) {
         // if has video, then audio buffer should not block the video buffer (bufferValue == 1, modified in AVDemuxThread)
         bv = statistics.audio.frame_rate > 0 && statistics.audio.frame_rate < 60 ?
-                        statistics.audio.frame_rate : 1;
+                        statistics.audio.frame_rate : 1LL;
     }
     if (buffer_mode == BufferTime)
-        bv = 600; //ms
+        bv = 600LL; //ms
     else if (buffer_mode == BufferBytes)
-        bv = 1024;
+        bv = 1024LL;
     // no block for music with cover
     if (video) {
         if (demuxer.hasAttacedPicture() || (statistics.video.frames > 0 && statistics.video.frames < bv))
-            bv = qMax<int>(1, statistics.video.frames);
+            bv = qMax<qint64>(1LL, statistics.video.frames);
     }
     buf->setBufferMode(buffer_mode);
-    buf->setBufferValue(buffer_value < 0 ? bv : buffer_value);
+    buf->setBufferValue(buffer_value < 0LL ? bv : buffer_value);
 }
 
 void AVPlayer::Private::updateBufferValue()
